Remove old dqn module
