+++ luchador/agent/rl/__init__.py
@@ -1,0 +1,4 @@
+"""Import Q Learning algorithms"""
+from __future__ import absolute_import
+
+from .q_learning import DeepQLearning  # noqa: F401
