+++ luchador/nn/tensorflow/__init__.py
@@ -8,7 +8,6 @@ from . import (  # noqa: F401
     layer,
     cost,
     optimizer,
-    q_learning,
 )
 from .scope import (  # noqa: F401
     VariableScope,
